Refactoring DLBDialogueDescription
Removed the com.dialoguebranch.parser.DLBFileDescription and replaced with the almost identical (old) com.dialoguebranch.model.DLBDialogueDescription (which is now called DLBFileDescription). These two objects where almost identical and the parser was doing quite confusing and useless conversions between the two.

The new DLBFileDescription also contains a "DLBFileType" parameter that indicates whether this is a Script or a Translation file.